SCRUM 26 - implement admin login functionality.
Updated the login.py file to check the role enum that is assigned to this user to check if they are a student or an admin role.

Updated the App.jsx file to include a route for the AdminDashboard for routing purposes.

Updated Login.jsx file to first check if user being logged in is a student or admin to determine either being sent to /home or /admin route.

Created AdminDashboard.jsx blank page to be routed to after admin logs in, still need to implement it more.